Move read_uci_id to a src folder and update download_data.py to reflect the change
