+++ src/read_uci_id.py
@@ -1,0 +1,46 @@
+from ucimlrepo import fetch_ucirepo 
+import pandas
+
+def read_uci_id(id, directory):
+    """
+    Read in a data set from the UCI Machine Learning repository using their API and save the 
+    contents to a specified directory.
+
+    Parameters:
+    -----------
+    id : int
+        The ucimlrepo id of the dataset to read in. 
+    directory : str
+        The directory where the data set will be saved.
+
+    Returns:
+    --------
+    None
+    """
+
+    # Create filename
+    filename_data = "raw_data.csv"
+    filename_data_sample = "raw_data_sample.csv"
+
+    # Cast id to int
+    id = int(id)
+
+    # Check if directory exists, if not raise error
+    if not os.path.isdir(directory):
+        raise ValueError('The directory provided does not exist.')
+
+    # Fetch the data from the UCI ML repo
+    uci_data = fetch_ucirepo(id=id)
+
+    # Create dataset and take a sample
+    raw_uci_data=uci_data.data.features; raw_uci_data['y']=uci_data.data.targets 
+    raw_uci_data_sample = raw_uci_data.sample(4000, random_state=522)
+
+    # Build path
+    full_path_data = os.path.join(directory, filename_data)
+    full_path_data_sample = os.path.join(directory, filename_data_sample)
+    
+    # write data to directory
+    raw_uci_data.to_csv(full_path_data)
+    raw_uci_data_sample.to_csv(full_path_data_sample)
+
